Consolidate a couple of methods into PeerInfo
