+++ lib/ruby_smb/client/authentication.rb
@@ -1,7 +1,11 @@
+require 'ruby_smb/peer_info'
+
 module RubySMB
   class Client
     # This module holds all the backend client methods for authentication.
     module Authentication
+      include RubySMB::PeerInfo
+
       # Responsible for handling Authentication and Session Setup for
       # the SMB Client. It returns the final Status code from the authentication
       # exchange.
@@ -355,36 +359,6 @@ module RubySMB
         packet.set_type3_blob(type3_message.serialize)
         packet.security_mode.signing_enabled = 1
         packet
-      end
-
-      # Extract and store useful information about the peer/server from the
-      # NTLM Type 2 (challenge) TargetInfo fields.
-      #
-      # @param target_info_str [String] the Target Info string
-      def store_target_info(target_info_str)
-        target_info = Net::NTLM::TargetInfo.new(target_info_str)
-        {
-          Net::NTLM::TargetInfo::MSV_AV_NB_COMPUTER_NAME  => :@default_name,
-          Net::NTLM::TargetInfo::MSV_AV_NB_DOMAIN_NAME    => :@default_domain,
-          Net::NTLM::TargetInfo::MSV_AV_DNS_COMPUTER_NAME => :@dns_host_name,
-          Net::NTLM::TargetInfo::MSV_AV_DNS_DOMAIN_NAME   => :@dns_domain_name,
-          Net::NTLM::TargetInfo::MSV_AV_DNS_TREE_NAME     => :@dns_tree_name
-        }.each do |constant, attribute|
-          if target_info.av_pairs[constant]
-            value = target_info.av_pairs[constant].dup
-            value.force_encoding('UTF-16LE')
-            instance_variable_set(attribute, value.encode('UTF-8'))
-          end
-        end
-      end
-
-      # Extract the peer/server version number from the NTLM Type 2 (challenge)
-      # Version field.
-      #
-      # @param version [String] the version number as a binary string
-      # @return [String] the formated version number (<major>.<minor>.<build>)
-      def extract_os_version(version)
-        version.unpack('CCS').join('.')
       end
     end
   end
